Added MPI and Threads Method
