--- build2/input.kmc
@@ -1,50 +1,0 @@
-#Build the lattice 
-lattice: SimpleCubic 100 100 10 A 
-
-#The growing film
-growth: CO2
-
-#Create steps (alaways in the x-direaction - see above 1st number), number of steps 
-#(e.g. if 10 each step will consist of two atoms), height difference (each step will increase by 20, so 1st step: 10, 2nd: 30, 3rd, 50 etc) 
-#steps: 2 1 
-
-#Total time in seconds
-time: 300
-
-#T in Kelvin
-temperature: 1000
-
-#P in Pascal
-pressure: 101325
-
-#Random number initialization
-random: 123434432
-
-#Simple s0*f*P/(2*pi*MW*Ctot*kb*T) -> Sticking coefficient [-], f [-], C_tot [sites/m2], MW [kg/mol] 
-#A + * -> A*: simple 0.1 2.0e-4 1.0e+19 0.032 
-
-#Constant is a constant value in [ML/s]
-CO + * -> CO*: constant 0.4 #0.389
-
-#Constant is a constant value in [ML/s]	
-O2 + 2* -> 2O*: constant 0.15 all #0.611
-		  		
-#A reaction 												
-CO* + O* -> CO2*: constant 1.e+15
-
-#Example of growth reaction. The 1sr reactant is tranformed to the 1st product, the 2nd reactan to the 2nd product etc. 
-#If #Reactants > #Products then only the first Nth reactants are tranformed to the (N-x) reactants. 
-#Products that have "*" stay on the lattice (adsorbed) 
-#products that do not have "*" are assumed to desorb instantly.  
-#CO* + O* -> CO2* : constant 0.25e+5
-
-
-#Time to write in log 
-write: log 0.1
-
-#Time to write the lattice heights & species
-write: lattice 10
-
-#Report the coverage of certain species. The time will follow the write in log file
-report: coverage CO* O*
-
